Add issue template, redirecting to swift-openapi-generator issues (#36)
### Motivation

We centralize the issues for all the repos in the Swift OpenAPI
Generator project in the generator repo. Using an issue template will
make this even clearer, because it will allow people to use the normal
Github workflow to discover the process and provide a link to where to
file their issue.

### Modifications

Add issue template, redirecting to swift-openapi-generator issues.

### Result

When people try and file an issue, they'll be presented with a button
that takes them to the generator repo issues page.

### Test Plan

Manual.